requirements.txt
aiohttp~=3.7.3
python-socketio~=5.0.2
chainchomplib~=0.0.10
asyncio